refactor(reducers):refactor reducers
- refactor post message to current group reducer
- combine all reducers about message into one file
- refactor message reducer test

[Finishes]
